flatten: speak_less_smile_more
