+++ packages/client/lib/app/core/widgets/speak_less_smile_more/speak_less_smile_more_store.dart
