--- packages/client/lib/app/core/widgets/speak_less_smile_more/stack/presentation/presentation.dart
@@ -1,2 +1,0 @@
-export 'mobx/speak_less_smile_more_store.dart';
-export 'widget/speak_less_smile_more.dart';
