--- packages/client/lib/app/core/widgets/speak_less_smile_more/stack/stack.dart
@@ -1,1 +1,0 @@
-export 'presentation/presentation.dart';
